+++ minilibft/minilibft.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   minilibft.h                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: moel-asr <moel-asr@student.1337.ma>        +#+  +:+       +#+        */
+/*   By: kmahdi <kmahdi@student.1337.ma>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/09 16:35:08 by moel-asr          #+#    #+#             */
-/*   Updated: 2023/07/08 17:05:16 by moel-asr         ###   ########.fr       */
+/*   Updated: 2023/07/09 00:32:20 by kmahdi           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,18 +16,18 @@
 # include <stdlib.h>
 # include <unistd.h>
 
-void	ft_bzero(void *s, size_t n);
-void	*ft_memcpy(void *dest, const void *src, size_t n);
-void	*ft_memmove(void *dest, const void *src, size_t len);
-void	*ft_memset(void *s, int c, size_t len);
-char	**ft_split(char const *s, char c);
-char	*ft_strchr(const char *str, int c);
-int		ft_strcmp(char *s1, char *s2);
-char	*ft_strdup(const char *str);
-char	*ft_strjoin(char const *s1, char const *s2);
-int		ft_strlen(const char *str);
-char	*ft_strstr(const char *str, const char *to_find);
-char	*ft_strtrim(char const *s1, char const *set);
-char	*ft_substr(char const *s, unsigned int start, size_t len);
-
-#endif
+void		ft_bzero(void *s, size_t n);
+void		*ft_memcpy(void *dest, const void *src, size_t n);
+void		*ft_memmove(void *dest, const void *src, size_t len);
+void		*ft_memset(void *s, int c, size_t len);
+char		**ft_split(char const *s, char c);
+char		*ft_strchr(const char *str, int c);
+int			ft_strcmp(char *s1, char *s2);
+char		*ft_strdup(const char *str);
+char		*ft_strjoin(char const *s1, char const *s2);
+int			ft_strlen(const char *str);
+char		*ft_strstr(const char *str, const char *to_find);
+char		*ft_strtrim(char const *s1, char const *set);
+char		*ft_substr(char const *s, int start, int len);
+	
+#endif
